refactor(suite): move suite sync turn-on ui into packages
