+++ packages/suite/src/components/suite/labeling/Labeling/Labeling.tsx
@@ -7,7 +7,7 @@ import {
     selectIsLabelingAvailableForEntity,
     selectMetadata,
 } from '@suite/metadata';
-import { SuiteSyncInteractionsTooltip } from '@suite/suite-sync';
+import { SuiteSyncInteractionsTooltip, TurnOnSuiteSyncModals } from '@suite/suite-sync';
 import { type MetadataAddPayload } from '@suite-common/metadata-types';
 import { selectIsSuiteSyncEnabled } from '@suite-common/suite-sync';
 import { type StaticSessionId } from '@trezor/connect';
@@ -19,7 +19,6 @@ import { useDiscovery, useDispatch, useSelector } from 'src/hooks/suite';
 import { useSuiteServices } from 'src/support/SuiteServicesProvider';
 
 import { selectIsLabelActionEnabled } from './selectIsLabelActionEnabled';
-import { TurnOnSuiteSyncModals } from '../TurnOnSuiteSync/TurnOnSuiteSyncModals';
 import { suiteSyncErrorHandler } from '../suiteSyncErrorHandler';
 
 type LabelingProps = {
@@ -166,6 +165,7 @@ export const Labeling = ({
                     onClose={() => handleSuiteSyncTurnOnModalComplete(false)}
                     onSuccess={() => handleSuiteSyncTurnOnModalComplete(true)}
                     deviceStaticSessionId={deviceStaticSessionId}
+                    suiteSync={suiteSync}
                 />
             )}
             <SuiteSyncInteractionsTooltip
