+++ packages/suite/src/components/suite/layouts/SuiteLayout/PassphraseFlow.tsx
@@ -1,5 +1,6 @@
 import { MODAL_CONTEXT_DEVICE, type MODAL_CONTEXT_NONE } from '@suite/modal';
 import { closeModalApp } from '@suite/router';
+import { TurnOnSuiteSyncModals } from '@suite/suite-sync';
 import { selectSelectedDevice } from '@suite-common/device';
 import { UI_REQUEST } from '@trezor/connect';
 
@@ -9,10 +10,10 @@ import {
 } from 'src/actions/suiteSync/suiteSyncSlice';
 import { ThpGlobalModalManager } from 'src/components/connection/thp/ThpGlobalModalManager';
 import { useDispatch, usePreferredModal, useSelector } from 'src/hooks/suite';
+import { useSuiteServices } from 'src/support/SuiteServicesProvider';
 import type { AppState, ForegroundAppRoute } from 'src/types/suite';
 import { SwitchDevice } from 'src/views/suite/SwitchDevice/SwitchDevice';
 
-import { TurnOnSuiteSyncModals } from '../../labeling/TurnOnSuiteSync/TurnOnSuiteSyncModals';
 import { ConfirmPassphraseBeforeAction } from '../../modals/ReduxModal/DeviceContextModal/ConfirmPassphraseBeforeAction';
 import { PassphraseModal } from '../../modals/ReduxModal/DeviceContextModal/PassphraseModal';
 import { PassphraseOnDeviceModal } from '../../modals/ReduxModal/DeviceContextModal/PassphraseOnDeviceModal';
@@ -59,6 +60,7 @@ type ForegroundAppModalProps = {
 
 const ForegroundAppModal = ({ app, cancelable }: ForegroundAppModalProps) => {
     const dispatch = useDispatch();
+    const { suiteSync } = useSuiteServices();
     const deviceStaticSessionId = useSelector(selectShowEnableSuiteSyncModal);
 
     const onCancel = () => dispatch(closeModalApp());
@@ -71,6 +73,7 @@ const ForegroundAppModal = ({ app, cancelable }: ForegroundAppModalProps) => {
                 <SwitchDevice cancelable={cancelable} onCancel={onCancel} />
                 <TurnOnSuiteSyncModals
                     deviceStaticSessionId={deviceStaticSessionId}
+                    suiteSync={suiteSync}
                     onClose={() => {
                         dispatch(updateShowEnableSuiteSyncModal({ deviceStaticSessionId: null }));
                     }}
