+++ suite/metadata-migration/src/index.ts
@@ -6,3 +6,4 @@ export {
     LegacyLabelingMigration,
     type LegacyLabelingMigrationProps,
 } from './LegacyLabelingMigration';
+export { SuiteSyncPromoBanner } from './SuiteSyncPromoBanner';
