+++ suite/suite-sync/package.json
@@ -17,6 +17,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/delegated-identity-key-types": "workspace:*",
         "@suite-common/device": "workspace:*",
+        "@suite-common/message-system": "workspace:*",
         "@suite-common/platform-encryption": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-sync-evolu": "workspace:*",
@@ -27,6 +28,7 @@
         "@suite-common/toast-notifications": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
         "@suite/analytics": "workspace:*",
+        "@suite/firmware-upgrade": "workspace:*",
         "@suite/intl": "workspace:*",
         "@trezor/components": "workspace:*",
         "@trezor/connect": "workspace:*",
