+++ suite/suite-sync/src/index.ts
@@ -1,6 +1,10 @@
 export { createSuiteSyncDesktopCompositionRoot } from './createSuiteSyncDesktopCompositionRoot';
 export { SelectSuiteSyncServer } from './SelectSuiteSyncServer';
 export { SuiteSyncInteractionsTooltip } from './SuiteSyncInteractionsTooltip';
+export { SuiteSyncTurnOnModal } from './SuiteSyncTurnOnModal';
+export { SuiteSyncTurnOnUnsupportedModal } from './SuiteSyncTurnOnUnsupportedModal';
 export { SuiteSyncServers } from './SuiteSyncServers';
 export { SuiteSyncSettings } from './SuiteSyncSettings';
+export { TurnOnSuiteSyncModals } from './TurnOnSuiteSyncModals';
 export { SuiteSyncWalletDebug } from './SuiteSyncWalletDebug';
+export { suiteSyncErrorTranslationKeyMap } from './suiteSyncErrorTranslationKeyMap';
